requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
plotly==5.17.0
requests==2.31.0
python-binance==1.0.17
textblob==0.17.1
vaderSentiment==3.3.2
python-dotenv==1.0.0